CM-019: Move the solution to another folder
