--- csharp-basic-topics/WriteXML/Tests/GlobalUsings.cs
@@ -1,6 +1,0 @@
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using System.Xml;
-global using System.Xml.Linq;
-global using WriteXML;
-global using WriteXML.Create;
-
