--- csharp-basic-topics/WriteXML/WriteXML/Resources/names.csv
@@ -1,21 +1,0 @@
-Name,Email,Telephone,Age
-John Doe,johndoe@email.com,555-1234,30
-Jane Smith,janesmith@email.com,555-5678,25
-Bob Johnson,bobjohnson@email.com,555-9876,35
-Alice Williams,alicewilliams@email.com,555-4321,28
-Chris Davis,chrisdavis@email.com,555-8765,40
-Emily White,emilywhite@email.com,555-2109,22
-Michael Lee,michaellee@email.com,555-6543,33
-Sara Brown,sarabrown@email.com,555-7890,29
-Daniel Turner,danielturner@email.com,555-3456,37
-Olivia Harris,oliviaharris@email.com,555-9012,26
-Ryan Miller,ryanmiller@email.com,555-2345,31
-Emma Martinez,emmamartinez@email.com,555-6789,27
-Jacob Carter,jacobcarter@email.com,555-4321,34
-Sophia Scott,sophiascott@email.com,555-8765,24
-Matthew Hall,matthewhall@email.com,555-2109,39
-Ava Wright,avawright@email.com,555-6543,23
-William Turner,williamturner@email.com,555-7890,32
-Grace Garcia,gracegarcia@email.com,555-3456,36
-Ethan Clark,ethanclark@email.com,555-9012,28
-Madison Adams,madisonadams@email.com,555-2345,33
