--- csharp-basic-topics/WriteXML/WriteXML/Resources/teams.csv
@@ -1,20 +1,0 @@
-Manchester United,Manchester
-Liverpool,Liverpool
-Chelsea,London
-Arsenal,London
-Manchester City,Manchester
-Tottenham Hotspur,London
-Leicester City,Leicester
-Everton,Liverpool
-West Ham United,London
-Aston Villa,Birmingham
-Newcastle United,Newcastle
-Southampton,Southampton
-Crystal Palace,London
-Wolverhampton Wanderers,Wolverhampton
-Leeds United,Leeds
-Brighton & Hove Albion,Brighton
-Burnley,Burnley
-Sheffield United,Sheffield
-Fulham,London
-West Bromwich Albion,West Bromwich
