+++ xml-csharp/WriteXML/Tests/GlobalUsings.cs
@@ -1,0 +1,6 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using System.Xml;
+global using System.Xml.Linq;
+global using WriteXML;
+global using WriteXML.Create;
+
